backend/requirements.txt
# Python Dependencies for Forgotten Ruin MUD

# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Channels for WebSocket support
channels==4.0.0
channels-redis==4.1.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Async support
daphne==4.0.0  # ASGI server

# Task Queue
celery==5.3.4
redis==5.0.1

# Password hashing
argon2-cffi==23.1.0

# CORS
django-cors-headers==4.3.0

# Utilities
python-dotenv==1.0.0  # Environment variables

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1

# Code Quality
flake8==6.1.0
black==23.11.0

# Security
safety==2.3.5  # Check for security vulnerabilities
